backEnd/bigdageeta/requirements.txt
alembic==1.4.2
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-mysql==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
Flask==1.1.2
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.1
idna==2.9
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.13.6
matplotlib==3.3.3
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.19.3
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
Pillow==8.0.1
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.0
scipy==1.5.4
six==1.14.0
sklearn==0.0
SQLAlchemy==1.2.13
sqlparse==0.3.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
